+++ tmo_monitor/gateway/model.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+class GatewayModel(Enum):
+  NOKIA = 'NOK5G21'
+  ARCADYAN = 'ARCKVD21'
